+++ packages/editor/web/src/RichContentEditor/handleBackspaceCommand.js
@@ -1,0 +1,37 @@
+import { RichUtils, EditorState } from '@wix/draft-js';
+import { isAtomicBlockFocused, removeBlock } from 'wix-rich-content-common';
+import removeBlockAdjacentToAtomic from './atomicBlockRemovalUtil';
+
+export default editorState => {
+  const selection = editorState.getSelection();
+
+  if (isAtomicBlockFocused(editorState)) {
+    return removeBlock(editorState, selection.getAnchorKey());
+  }
+
+  if (!selection.isCollapsed() || selection.getAnchorOffset() || selection.getFocusOffset()) {
+    return null;
+  }
+
+  // First, try to remove a preceding atomic block.
+  const content = editorState.getCurrentContent();
+  const startKey = selection.getStartKey();
+  const blockBefore = content.getBlockBefore(startKey);
+
+  if (blockBefore && blockBefore.getType() === 'atomic') {
+    const withoutCurrentBlock = removeBlockAdjacentToAtomic(editorState, false);
+
+    if (withoutCurrentBlock) {
+      return withoutCurrentBlock;
+    }
+  }
+
+  // If that doesn't succeed, try to remove the current block style.
+  const withoutBlockStyle = RichUtils.tryToRemoveBlockStyle(editorState);
+
+  if (withoutBlockStyle) {
+    return EditorState.push(editorState, withoutBlockStyle, 'change-block-type');
+  }
+
+  return null;
+};